Fixed some errors. Differentiated betweeen pan id and node id. Prepare NVM for node id of 64 bits. Disable firmware update in configuration of the nodes.

+++ header/common.h
@@ -33,7 +33,7 @@
 
 typedef struct {
   int cnt;
-  uint16_t node_ids[DWM_RANGING_ANCHOR_CNT_MAX];
+  uint64_t node_ids[DWM_RANGING_ANCHOR_CNT_MAX];
 } rangin_neighbors;
 
 /*************
